% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-icd.R
\name{add_dot}
\alias{add_dot}
\title{Add a dot to an ICD-10 code}
\usage{
add_dot(x)
}
\arguments{
\item{x}{A valid ICD-10 code without a dot}
}
\value{
A valid ICD-10 code with a dot included
}
\description{
\code{add_dot()} adds a dot to the ICD-10 code in the appropriate position
where one does not exist
}
\examples{
add_dot("F320")

add_dot("F32")  # no dot added if code is only 3-digits
}
