% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendixA.R
\name{appendix_A}
\alias{appendix_A}
\title{Appendix A: List of MS-DRGs Version 41.1}
\usage{
appendix_A(drg = NULL, mdc = NULL, type = NULL, ...)
}
\arguments{
\item{drg}{\verb{<chr>} vector of 3-digit DRG codes}

\item{mdc}{\verb{<chr>} vector of 2-digit Major Diagnostic Category codes}

\item{type}{\verb{<chr>} DRG type: \code{M} (Medical) or \code{P} (Surgical)}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Appendix A contains a list of each MS-DRG with a specification of the MDC and
whether the MS-DRG is medical or surgical. Some MS-DRGs which contain patients
from multiple MDCs (e.g., 014 Allogeneic Bone Marrow Transplant) do not have
an MDC specified. The letter M is used to designate a medical MS-DRG and the
letter P is used to designate a surgical MS-DRG.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
appendix_A(drg = "011")

appendix_A(mdc = "24")
\dontshow{\}) # examplesIf}
}
