% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendixB.R
\name{appendix_B}
\alias{appendix_B}
\title{Appendix B: Diagnosis Code/MDC/MS-DRG Index}
\usage{
appendix_B()
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
The Diagnosis Code/MDC/MS-DRG Index lists each diagnosis code, as well as
the MDC, and the MS-DRGs to which the diagnosis is used to define the logic
of the DRG either as a principal or secondary diagnosis.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
appendix_B()
\dontshow{\}) # examplesIf}
}
