% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{appendix_G}
\alias{appendix_G}
\title{Appendix G Diagnoses Defined as Complications or Comorbidities}
\usage{
appendix_G()
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Diagnoses in Appendix G are considered complications or comorbidities (CC)
except when used in conjunction with the principal diagnosis in the
corresponding CC Exclusion List in Appendix C.
}
\examples{
head(appendix_G())

}
