% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{appendix_H}
\alias{appendix_H}
\title{Appendix H Diagnoses Defined as Major Complications or Comorbidities}
\usage{
appendix_H()
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Diagnoses in Appendix H Part I are considered major complications or
comorbidities (MCC) except when used in conjunction with the principal
diagnosis in the corresponding CC Exclusion List in Appendix C. In addition to
the CC exclusion list, the diagnoses in Part II are assigned as a major CC
only for patients discharged alive, otherwise they will be assigned as a non
CC.
}
\examples{
head(appendix_H())

}
