% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_edits.R
\name{apply_age_edits}
\alias{apply_age_edits}
\title{Apply Age Conflict Edits}
\usage{
apply_age_edits(rule, age)
}
\arguments{
\item{rule}{\verb{<chr>} ICD-10-CM Conflict Rule}

\item{age}{\verb{<int>} age in years}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Apply Age Conflict Edits
}
\examples{
apply_age_edits(
   c("Pediatric (Ages 0-17)",
     "Maternity (Ages 9-64)"),
   c(18, 7))

}
