% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-icd.R
\name{case_chapter_icd10}
\alias{case_chapter_icd10}
\title{Add ICD-10-CM Chapter Labels}
\usage{
case_chapter_icd10(df, col)
}
\arguments{
\item{df}{\verb{<chr>}data frame}

\item{col}{\verb{<sym>} unquoted column name of ICD-10-CM codes to match}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{chapter} column
}
\description{
Add ICD-10-CM Chapter Labels
}
\examples{
dplyr::tibble(code = c(
              "F50.8", "G40.311", "Q96.8", "Z62.890", "R45.4",
              "E06.3", "H00.019", "D50.1", "C4A.70", "Z20.818")) |>
              case_chapter_icd10(code)
}
