% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_marks}
\alias{gt_marks}
\title{gt Marks}
\usage{
gt_marks(tbl, cols)
}
\arguments{
\item{tbl}{\verb{<gt_tbl>} object}

\item{cols}{\verb{<syms>} vector of bare column names, e.g. \code{c(mdc, drg)}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
gt Marks
}
\keyword{internal}
