% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icd_code_to_chapter}
\alias{icd_code_to_chapter}
\title{Match ICD-10-CM Codes to Chapter Labels}
\usage{
icd_code_to_chapter(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of ICD-10-CM codes}
}
\value{
\verb{<chr>} vector of ICD-10-CM chapter labels
}
\description{
Match ICD-10-CM Codes to Chapter Labels
}
\examples{
codes <- c("F50.8", "G40.311", "Q96.8",
           "R45.4", "E06.3", "H00.019",
           "D50.1", "C4A.70", "Z20.818")

icd_code_to_chapter(codes)

data.frame(code = codes,
           chapter = icd_code_to_chapter(codes))

}
