% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_period}
\alias{insert_period}
\title{Insert period into an ICD-10-CM code}
\usage{
insert_period(x)
}
\arguments{
\item{x}{vector of ICD-10-CM codes}
}
\value{
vector of ICD-10-CM codes with period inserted
}
\description{
Inserts period into an ICD-10-CM code in between the 3rd and 4th
characters, if not present or if the code is longer than 3 characters
}
\examples{
insert_period("F320")

# Not inserted if code is only 3 characters
insert_period("F32")

}
