% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_icd}
\alias{is_valid_icd}
\title{Check if string is valid ICD-10-CM code}
\usage{
is_valid_icd(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{\verb{<chr>} string of \code{length(1)}}

\item{arg}{\verb{<chr>} function argument name in the current function}

\item{call}{\verb{<environment>} environment the function is called from}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, \code{FALSE} otherwise
}
\description{
Check if string is valid ICD-10-CM code
}
\examples{
purrr::map_vec(
c("H00.019", "D50.1", "C4A.70", "Z20.818", "11646", "E8015"),
is_valid_icd)

}
