% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_icd10cm}
\alias{is_valid_icd10cm}
\title{Check if an ICD-10-CM code is in a valid format}
\usage{
is_valid_icd10cm(x)
}
\arguments{
\item{x}{A string}
}
\value{
Boolean
}
\description{
\code{is_valid_icd10cm} checks to see if the ICD-10-CM code is valid
and has between 3 and 8 characters, and starts with a letter and number.
}
\examples{
is_valid_icd10cm("F320") # valid
is_valid_icd10cm("F32")  # valid
is_valid_icd10cm("296")  # invalid

}
