% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mount_board}
\alias{mount_board}
\title{Mount \link[pins:pins-package]{pins} board}
\usage{
mount_board(source = c("local", "remote"))
}
\arguments{
\item{`<chr>`}{string; whether source is \code{"local"} or \code{"remote"}}
}
\value{
\verb{<pins_board_folder>} if \code{source = "local"} or \verb{<pins_board_url>}
if \code{source = "remote"}
}
\description{
Mount \link[pins:pins-package]{pins} board
}
\keyword{internal}
