% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{msdrg_index}
\alias{msdrg_index}
\title{ICD-10-CM|MDC|MS-DRG Index}
\usage{
msdrg_index(icd = NULL, drg = NULL, mdc = NULL)
}
\arguments{
\item{icd}{\verb{<chr>} ICD-10-CM code}

\item{drg}{\verb{<chr>} vector of 3-digit Diagnosis-Related Group codes}

\item{mdc}{\verb{<chr>} vector of 2-digit Major Diagnostic Category codes}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\strong{Appendix B}: ICD-10-CM|MDC|MS-DRG Index
}
\details{
The Diagnosis Code/MDC/MS-DRG Index lists each ICD-10-CM code,
the MDC and the MS-DRGs to which the diagnosis is used to define
the logic of the MS-DRG either as a principal or secondary diagnosis.
}
\examples{
msdrg_index(icd = "A17.81")

msdrg_index(drg = "011")

msdrg_index(mdc = "24")

}
