% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{msdrg_list}
\alias{msdrg_list}
\title{MS-DRG List}
\usage{
msdrg_list(drg = NULL, mdc = NULL, type = NULL)
}
\arguments{
\item{drg}{\verb{<chr>} vector of 3-digit Diagnosis-Related Group codes}

\item{mdc}{\verb{<chr>} vector of 2-digit Major Diagnostic Category codes}

\item{type}{\verb{<chr>} \code{"M"} (Medical) or \code{"P"} (Surgical)}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\strong{Appendix A}: Contains each MS-DRG with a specification of
the MDC and whether the MS-DRG is Medical or Surgical.
}
\details{
Some MS-DRGs which contain patients from multiple MDCs
e.g., \code{"014"} (Allogeneic Bone Marrow Transplant) do not
have an MDC specified.

The letter M is used to designate a medical MS-DRG and the
letter P is used to designate a surgical MS-DRG.
}
\examples{
msdrg_list(drg = "014")

msdrg_list(mdc = "24")

msdrg_list(type = "M")

}
