% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_icd}
\alias{random_icd}
\title{Generate Random ICD-10-CM Codes}
\usage{
random_icd(n = 10, lng = c(3, 5:8))
}
\arguments{
\item{n}{\verb{<numeric>} number of codes; default is \code{10}}

\item{lng}{\verb{<numeric>} vector of lengths; default is \code{c(3, 5:8)}}
}
\value{
\verb{<character>} vector
}
\description{
Generate Random ICD-10-CM Codes
}
\examples{
random_icd()

random_icd(5, 8)

}
