% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_period}
\alias{remove_period}
\title{Remove period from ICD-10-CM code}
\usage{
remove_period(x)
}
\arguments{
\item{x}{A valid ICD-10 code with a dot}
}
\value{
A valid ICD-10 code without a dot included
}
\description{
Removes period from the ICD-10 code if it exists
}
\examples{
remove_period("F32.0")

remove_period("F32")

}
