% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdrg.R
\name{search_edits}
\alias{search_edits}
\title{Search ICD-10-CM Code Edits}
\usage{
search_edits(code = NULL, ...)
}
\arguments{
\item{code}{\verb{<chr>} vector of ICD-10-CM codes}

\item{...}{Empty dots}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Definitions of Medicare Code Edits, version 41.1
}
\examples{
search_edits(c("Q96.8", "N47.0", "R45.4", "A33"))

}
