% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_in}
\alias{search_in}
\title{Search in data frame}
\usage{
search_in(df, dfcol, search)
}
\arguments{
\item{df}{\verb{<data.frame>}}

\item{dfcol}{\verb{<sym>} unquoted column to search, in the form of \code{df$col}}

\item{search}{\verb{<chr>} vector of strings to search for in \code{df$col}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search in data frame
}
\keyword{internal}
