% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdrg.R
\name{search_msdrg}
\alias{search_msdrg}
\title{Medicare Severity Diagnosis-Related Groups}
\usage{
search_msdrg(drg = NULL, mdc = NULL, type = NULL, ...)
}
\arguments{
\item{drg}{\verb{<chr>} vector of 3-digit DRG codes}

\item{mdc}{\verb{<chr>} vector of 2-digit Major Diagnostic Category codes}

\item{type}{\verb{<chr>} DRG type: \code{Medical} or \code{Surgical}}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
The Medicare Severity Diagnosis-Related Group (MS-DRG) is a classification
system used by the Centers for Medicare and Medicaid Services (CMS) to group
patients with similar clinical characteristics and resource utilization into
a single payment category.
}
\details{
The system is primarily used for Medicare reimbursement purposes, but it is
also adopted by many other payers as a basis for payment determination.

MS-DRGs are based on the principal diagnosis, up to 24 additional diagnoses,
and up to 25 procedures performed during the stay. In a small number of
MS-DRGs, classification is also based on the age, sex, and discharge status
of the patient.

Hospitals serving more severely ill patients receive increased
reimbursements, while hospitals treating less severely ill patients will
receive less reimbursement.
}
\examples{
search_msdrg(drg = "011")

search_msdrg(mdc = "24")
}
