% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_edits.R
\name{years_floor}
\alias{years_floor}
\title{Calculate the number of years between two dates}
\usage{
years_floor(from, to)
}
\arguments{
\item{from}{\verb{<date>} Start date}

\item{to}{\verb{<date>} End date}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculate the number of years between two dates
}
\examples{
years_floor(
   from = as.Date("2020-01-01"),
   to   = as.Date("2020-01-01") + 2057)

}
