% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{appendix_C}
\alias{appendix_C}
\title{Appendix C: Complications or Comorbidities Exclusion list}
\usage{
appendix_C()
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Appendix C is a list of all of the codes that are defined as either a
complication or comorbidity (CC) or a major complication or comorbidity (MCC)
when used as a secondary diagnosis.
}
\details{
Part 1 lists these codes. Each code is indicated as CC or MCC. A link is given
to a collection of diagnosis codes which, when used as the principal
diagnosis, will cause the CC or MCC to be considered as only a non-CC.

Part 2 lists codes which are assigned as a Major CC only for patients
discharged alive. Otherwise they are assigned as a non-CC.

Part 3 lists diagnosis codes that are designated as a complication or
comorbidity (CC) or major complication or comorbidity (MCC) and included in
the definition of the logic for the listed DRGs. When reported as a secondary
diagnosis and grouped to one of the listed DRGs the diagnosis is excluded from
acting as a CC/MCC for severity in DRG assignment.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
appendix_C()
\dontshow{\}) # examplesIf}
}
