% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_long}
\alias{display_long}
\title{Pivot data frame to long format for easy printing}
\usage{
display_long(df, cols = dplyr::everything())
}
\arguments{
\item{df}{\verb{<data.frame>}}

\item{cols}{\verb{<syms>} vector of bare column name to pivot, default is \code{\link[dplyr:reexports]{dplyr::everything()}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Pivot data frame to long format for easy printing
}
\keyword{internal}
