% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_marks}
\alias{gt_marks}
\title{gt Marks}
\usage{
gt_marks(gt_tbl, cols)
}
\arguments{
\item{gt_tbl}{description}

\item{cols}{description}
}
\value{
description
}
\description{
gt Marks
}
\keyword{internal}
