% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_style}
\alias{gt_style}
\title{gt Theme}
\usage{
gt_style(tbl, lbl = TRUE, tablign = "center", tabsize = 16, tabwt = "normal")
}
\arguments{
\item{tbl}{\verb{<gt_tbl>} object}

\item{lbl}{\verb{<lgl>} hide column labels; default is \code{TRUE}}

\item{tablign}{\verb{<chr>} table stub alignment; default is \code{"center"}}

\item{tabsize}{\verb{<int>} table stub font size in pixels; default is \code{16}}

\item{tabwt}{\verb{<chr>} table stub font weight; default is \code{"normal"}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
gt Theme
}
\keyword{internal}
