% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_api.R
\name{icd10api}
\alias{icd10api}
\title{National Library of Medicine's ICD-10-CM API}
\source{
National Library of Medicine/National Institute of Health
}
\usage{
icd10api(
  code = NULL,
  term = NULL,
  field = c("both", "code"),
  limit = 500L,
  ...
)
}
\arguments{
\item{code}{\verb{<chr>} ICD-10-CM code}

\item{term}{\verb{<chr>} Associated term describing an ICD-10 code}

\item{field}{\verb{<chr>} \code{code} or \code{both}; default is \code{both}}

\item{limit}{\verb{<int>} API limit, defaults to 500}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=icd10api]{icd10api()}} allows you to search the National Library of
Medicine's ICD-10-CM API by code or associated term.
}
\details{
ICD-10-CM (International Classification of Diseases, 10th Revision,
Clinical Modification) is a medical coding system for classifying
diagnoses and reasons for visits in U.S. health care settings.
\subsection{Links}{
\itemize{
\item \href{https://clinicaltables.nlm.nih.gov/apidoc/icd10cm/v3/doc.html}{NIH NLM Clinical Table Service ICD-10-CM API}
\item \href{http://www.cdc.gov/nchs/icd/icd10cm.htm}{Learn more about ICD-10-CM.}
}
}
}
\note{
Current Version: ICD-10-CM \strong{2024}
}
\examples{
# Returns the seven codes
# beginning with "A15"
icd10api(code = "A15")

# Returns the first five codes
# associated with tuberculosis
icd10api(term = "tuber", limit = 5)

# Returns the two codes
# associated with pleurisy
icd10api(term = "pleurisy")

# If you're searching for codes beginning
# with a certain letter, you must set the
# `field` param to `"code"` or it will
# search for terms as well:

# Returns terms containing the letter "Z"
icd10api(code = "z", limit = 5)

# Returns codes beginning with "Z"
icd10api(code = "z", field = "code", limit = 5)
}
