% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathologie-package.R
\docType{package}
\name{pathologie-package}
\alias{pathologie}
\alias{pathologie-package}
\title{pathologie: Tidy ICD-10-CM Tools}
\description{
Tools for working with ICD-10-CM codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/pathologie/}
  \item \url{https://github.com/andrewallenbruce/pathologie}
  \item Report bugs at \url{https://github.com/andrewallenbruce/pathologie/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrew.bruce@northstarrevenueintegrity.com} [copyright holder]

}
\keyword{internal}
