% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathologie-package.R
\docType{package}
\name{pathologie-package}
\alias{pathologie}
\alias{pathologie-package}
\title{pathologie: Tidy ICD-10-CM Tools}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrew.bruce@northstarrevenueintegrity.com} [copyright holder]

}
\keyword{internal}
