% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_edits.R
\name{search_edits}
\alias{search_edits}
\title{ICD-10-CM Code Edits}
\usage{
search_edits(icd = NULL, group = NULL)
}
\arguments{
\item{icd}{\verb{<chr>} vector of ICD-10-CM codes}

\item{group}{\verb{<chr>} Conflict Group: \code{Age}, \code{Sex}, \code{Other}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Definitions of Medicare Code Edits, version 41.1
}
\details{
FY 2024 – Version 41.1 (Effective April 1, 2024 through September 30, 2024)

The \strong{ICD-10 Definitions of Medicare Code Edits} file contains the following: A
description of each coding edit with the corresponding code lists as well as
all the edits and the code lists effective for FY 2024.
}
